Use the mosquitto-go-auth plugin that replaces mosquitto-auth-plugin.
Whilst they appear similar there are subtle differences; config options
are scoped to the backend type (ie auth_mysql_setting rather than
auth_setting), the rw permission numbers in the DB have changed with mosquitto 1.5
(to subscribe must be 4 instead of 2).

The current hash generation in the remoteaccess-server module does not work
with this plugin either (maybe settings or base64 padding?), but just use the
pw command that ships with the plugin.